depV to double, write_wp inside right if

+++ deps/src/mjd_siggen/mjd_siggen.h
@@ -129,7 +129,7 @@ typedef struct {
   double **vfraction, *s1, *s2, **vsave;
   char   **point_type, **undepleted;
   int    fully_depleted;
-  int    depV;
+  double depV;
   float  bubble_volts, Emin;
   float  rho_z_spe[1024];
   double **dr[2], **dz[2];
